Refactor framework directories
