+++ app/src/main/java/com/zeronfinity/cpfy/framework/di/ApplicationModule.kt
@@ -4,10 +4,8 @@ import android.app.Application
 import com.zeronfinity.core.repository.*
 import com.zeronfinity.core.usecase.GetCookieUseCase
 import com.zeronfinity.cpfy.CustomApplication
-import com.zeronfinity.cpfy.framework.network.clist.RetrofitClistApiClient
-import com.zeronfinity.cpfy.framework.network.clist.RetrofitClistApiInterface
-import com.zeronfinity.cpfy.model.*
-import com.zeronfinity.cpfy.model.network.ClistNetworkCall
+import com.zeronfinity.cpfy.framework.implementations.*
+import com.zeronfinity.cpfy.framework.network.ClistNetworkCall
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -17,23 +15,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(ApplicationComponent::class)
 class ApplicationModule {
-    @Singleton
-    @Provides
-    fun provideClistClient(
-        application: Application,
-        getCookieUseCase: GetCookieUseCase
-    ): RetrofitClistApiClient =
-        RetrofitClistApiClient(application as CustomApplication, getCookieUseCase)
-
-    @Singleton
-    @Provides
-    fun provideClistApi(retrofitClistApiClient: RetrofitClistApiClient): RetrofitClistApiInterface = retrofitClistApiClient.getClistApi()
-
-    @Singleton
-    @Provides
-    fun provideClistNetworkCall(application: Application, apiInterface: RetrofitClistApiInterface) =
-        ClistNetworkCall(application as CustomApplication, apiInterface)
-
     @Singleton
     @Provides
     fun provideContestRepository() = ContestRepository(ContestArrayList())
