+++ app/src/test/java/com/zeronfinity/cpfy/implementations/FilterTimeRangeSharedPreferencesTest.kt
@@ -1,7 +1,8 @@
-package com.zeronfinity.cpfy.model
+package com.zeronfinity.cpfy.implementations
 
 import android.app.Application
 import android.content.SharedPreferences
+import com.zeronfinity.cpfy.framework.implementations.FilterTimeRangeSharedPreferences
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
