+++ build.gradle
@@ -12,7 +12,7 @@ buildscript {
     ext.okhttp_version = "4.9.0"
     ext.picasso_version = "2.71828"
     ext.retrofit2_version = "2.9.0"
-    ext.room_version = "2.2.5"
+    ext.room_version = "2.2.6"
 
     repositories {
         google()
